requirements.txt
Django
python-dotenv
djangorestframework
drf-yasg
djangorestframework-simplejwt
django-filter
psycopg2-binary